Use pycodestyle in place of pep8
In Foxy and newer, this package has been replaced with
`ament_pycodestyle`. Since the Python API for pep8 is compatible with
pycodestyle, we can switch to pycodestyle under-the-hood so that
downstream packages don't need to be updated.

If downstream packages provide their own configuration, they will get a
deprecation warning about switching the section header to `pycodestyle`,
but I'm not aware of any packages which do this.

Signed-off-by: Scott K Logan <logans@cottsay.net>